stm32: Negative values in layer set now leaves some layers unchanged

+++ STM32/logic/logic_layers.h
@@ -32,7 +32,7 @@ unsigned int logic_layers_refresh_cnt();
 
 unsigned int logic_layers_max();
 
-void logic_layers_select(const uint8_t *layers, unsigned int start, unsigned int num);
+void logic_layers_select(const uint8_t *layers, int8_t start, unsigned int num);
 void logic_layers_set_param(unsigned int layer, const void *p, unsigned int size);
 void *logic_layers_active_obj(unsigned int layer, layer_obj_enum_t obj, unsigned int *size);
 void *logic_layers_inactive_obj(unsigned int layer, layer_obj_enum_t obj, unsigned int *size);
